requirements.txt
# ApexNeural tracking microservice — production deps (pinned where practical)
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Async PostgreSQL
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Alembic (migrations; uses sync driver for CLI)
alembic>=1.13.0
psycopg[binary]>=3.2.0
